requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Google APIs
google-generativeai>=0.3.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0

# Configuration and environment
pyyaml>=6.0.1
python-dotenv>=1.0.0

# PDF generation
reportlab>=4.0.0
weasyprint>=60.0

# HTML/Email
jinja2>=3.1.2
premailer>=3.10.0

# Database
# SQLite is built into Python

# Utilities
python-dateutil>=2.8.2
tenacity>=8.2.0
fake-useragent>=1.4.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.10.0
pylint>=3.0.0
mypy>=1.6.0

# Optional: Browser automation (uncomment if needed)
# selenium>=4.15.0
# playwright>=1.39.0
